% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{default_engine_project_dir}
\alias{default_engine_project_dir}
\title{Locate the bundled Julia project directory (containing Project.toml).}
\usage{
default_engine_project_dir()
}
\description{
Locate the bundled Julia project directory (containing Project.toml).
}
\keyword{internal}
