% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{default_engine_src_dir}
\alias{default_engine_src_dir}
\title{Locate the bundled Julia engine source directory.}
\usage{
default_engine_src_dir()
}
\description{
Locate the bundled Julia engine source directory.
}
\keyword{internal}
