% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{load_engine}
\alias{load_engine}
\title{Include every \verb{*.jl} file in \code{src_dir} into Main.}
\usage{
load_engine(src_dir)
}
\description{
Include every \verb{*.jl} file in \code{src_dir} into Main.
}
\keyword{internal}
